support extracting textures in the TX_START/TX_END namespace.
The textures are saved into the "tx_start" directory, and
remembered in wadinfo.txt in the [TX_START] section.

Lumps which are detected to be PNG or JPEG images are saved
as-is (no conversion is done), otherwise we assume they are
DOOM patch format and attempt to convert them to a real image
format (just like graphics, patches, etc are handled).