Updated entry parser to support a "/raw" flag.
This flag generally must appear after the entry name, but
may also appear after the "= filename" syntax (for consistency
with XY offsets which can appear in either place).

The beginning slash is part of the flag, and the parser could
easily be extended to recognise new flags, which may be useful
in the future.

Main use for this "/raw" flag is with the [TX_START] section,
where it causes images to be inserted directly into the wad
(the flag replaces the previously hacky method of using the XY
offsets to mark such images).  This "/raw" flag may have more
uses in the future.

I also renamed TXTentryParse --> TXTparseEntry, consistent
with the "verb + Noun" format of most other functions.
Plus I improved the way the "= filename" syntax was parsed.